+++ RELEASENOTES.md
@@ -10,6 +10,31 @@
 
 By setting failOn to 'newWarning', pull requests will fail if new warnings are introduced. This feature compares the warnings in the pull request build against those in the latest successful CI/CD build and fails if new warnings are detected.
 
+### New Workflow specific settings
+
+- `buildEnvironmentName` - should be the Business Central environment name (from the Admin Center) for the environment you want to use if you want to run tests using online environments. Note that you also need to create a [buildAuthContext](https://aka.ms/algosecrets#buildAuthContext) secret, which gives access to this environment.
+
+### Use online environment during build for running tests
+
+By creating a [build authentication context](https://aka.ms/algosecrets#buildAuthContext) secret and setting the [buildEnvironmentName](https://aka.ms/algosettings#buildEnvironmentName) for a specific branch and a specific workflow, will cause AL-Go for GitHub to use this environment for running tests. An example of how to define settings if you have dependencies to 3rd party AppSource apps, where you do not have runtime packages to satisfy the dependencies, could be:
+
+```json
+  "useCompilerFolder": true,
+  "doNotPublishApps": true,
+  "conditionalSettings": [
+    {
+      "workflows": [ "CICD" ],
+      "branches": [ "main" ],
+      "settings": {
+        "doNotPublishApps": false,
+        "buildEnvironmentName": "cicd",
+        "buildAuthContextSecretName": "cicdAuthContext",
+        "workflowConcurrency": [ "group: ${{ github.workflow }}-${{ github.ref }}", "cancel-in-progress: false" ]
+      }
+    }
+  ]
+```
+
 ## v7.1
 
 ### Issues
@@ -899,7 +924,7 @@ Setting the repo setting "runs-on" to "Ubuntu-latest", followed by running Updat
 - New setting: **DoNotPublishApps** - setting this to true causes the workflow to skip publishing, upgrading and testing the app to improve performance.
 - New setting: **ConditionalSettings** to allow to use different settings for specific branches. Example:
 
-```
+```json
     "ConditionalSettings": [
         {
             "branches": [